# Change parquet FindTraceByID to not use page index