--- a/tempodb/encoding/vparquet/block_findtracebyid_test.go
+++ b/tempodb/encoding/vparquet/block_findtracebyid_test.go
@@ -3,6 +3,7 @@ package vparquet
 import (
 	"bytes"
 	"context"
+	"path"
 	"sort"
 	"testing"
 
@@ -136,5 +137,29 @@ func TestBackendBlockFindTraceByID_TestData(t *testing.T) {
 		protoTr, err := b.FindTraceByID(ctx, tr.TraceID, common.SearchOptions{})
 		require.NoError(t, err)
 		require.NotNil(t, protoTr)
+	}
+}
+
+func BenchmarkFindTraceByID(b *testing.B) {
+	ctx := context.TODO()
+	tenantID := "1"
+	blockID := uuid.MustParse("3685ee3d-cbbf-4f36-bf28-93447a19dea6")
+
+	r, _, _, err := local.New(&local.Config{
+		Path: path.Join("/Users/marty/src/tmp/"),
+	})
+	require.NoError(b, err)
+
+	rr := backend.NewReader(r)
+
+	meta, err := rr.BlockMeta(ctx, blockID, tenantID)
+	require.NoError(b, err)
+
+	block := newBackendBlock(meta, rr)
+
+	for i := 0; i < b.N; i++ {
+		tr, err := block.FindTraceByID2(ctx, meta.MinID, defaultSearchOptions())
+		require.NoError(b, err)
+		require.NotNil(b, tr)
 	}
 }