**What this PR does**:
This PR introduces an alternate version of parquet.FindTraceByID which doesn't load the page index.  It still does a binary search over rowgroup min/max bounds to find the trace, but instead of using the bounds in the page index, it establishes the lower bound of each row group by reading a single trace ID from it.   We have found that with larger blocks (10-30GB) the page index can become quite large (250+MB) and costly to unmarshal.  Although this implementation will perform `log2(numRowGroups)` reads they are much smaller and overall faster and more efficient.

Of course, there is a breakpoint depending on the object store read latency and block size.  Below are benchmarks with three levels of simulated latency (time.Sleep in the ReaderAt) and two block sizes.  Above 100ms latency or smaller blocks, it becomes worse.  Not sure the explanation for the variable alloc/op after the change, it should be deterministic but perhaps it has to do with number of iterations per benchmark and reused buffers which get amortized.

PR is marked draft because I'd like to gather some thoughts on it, perhaps we could choose the approach dynamically based on the block. 

9GB block:
```
name              old time/op    new time/op    delta
FindTraceByID-12     539ms ± 1%     255ms ± 2%  -52.62%  (p=0.000 n=8+8)    20ms simulated latency
FindTraceByID-12     764ms ± 4%     589ms ± 2%  -22.98%  (p=0.000 n=9+10)   50ms simulated latency
FindTraceByID-12     1.11s ± 3%     1.13s ± 0%     ~     (p=0.059 n=8+9)   100ms simulated latency

name              old alloc/op   new alloc/op   delta
FindTraceByID-12     185MB ± 3%     115MB ±36%  -37.66%  (p=0.000 n=10+10)
FindTraceByID-12     184MB ± 2%     138MB ± 4%  -25.22%  (p=0.000 n=9+10)
FindTraceByID-12     188MB ± 2%      32MB ±20%  -82.72%  (p=0.000 n=9+9)

name              old allocs/op  new allocs/op  delta
FindTraceByID-12     2.06M ± 0%     0.07M ± 0%  -96.83%  (p=0.000 n=9+9)
FindTraceByID-12     2.06M ± 0%     0.07M ± 0%  -96.83%  (p=0.000 n=10+10)
FindTraceByID-12     2.06M ± 0%     0.07M ± 0%  -96.83%  (p=0.000 n=9+9)
```

700MB block
```
FindTraceByID-12     256ms ±28%     278ms ±12%     ~     (p=0.065 n=10+9)
FindTraceByID-12     453ms ± 2%     556ms ± 3%  +22.76%  (p=0.000 n=8+10)
FindTraceByID-12     862ms ± 1%    1063ms ± 3%  +23.22%  (p=0.000 n=10+8)
FindTraceByID-12     823MB ±16%     746MB ±20%     ~     (p=0.065 n=10+9)

name              old alloc/op   new alloc/op   delta
FindTraceByID-12     780MB ±12%     805MB ±15%     ~     (p=0.315 n=10+10)
FindTraceByID-12     780MB ±10%     735MB ±12%     ~     (p=0.063 n=10+10)

name              old allocs/op  new allocs/op  delta
FindTraceByID-12     29.9k ± 1%     16.2k ± 1%  -45.69%  (p=0.000 n=10+10)
FindTraceByID-12     30.0k ± 1%     16.3k ± 1%  -45.69%  (p=0.000 n=10+10)
FindTraceByID-12     30.1k ± 1%     16.3k ± 6%  -45.74%  (p=0.000 n=10+9)
```

**Which issue(s) this PR fixes**:
Fixes #<issue number>

**Checklist**
- [ ] Tests updated
- [ ] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`